Cleanup and use switch method
