--- src/Lynx.Benchmark/CastleHash.cs
@@ -1,74 +1,0 @@
-﻿/*
- *
-
- */
-
-using BenchmarkDotNet.Attributes;
-using Lynx.Model;
-using System.Runtime.CompilerServices;
-using static Lynx.ZobristTable;
-
-namespace Lynx.Benchmark;
-public class CastleHash : BaseBenchmark
-{
-    public static IEnumerable<Position> Data => new[] {
-        new Position(Constants.InitialPositionFEN),
-        new Position(Constants.TrickyTestPositionFEN),
-        new Position(Constants.TrickyTestPositionReversedFEN),
-        new Position(Constants.CmkTestPositionFEN),
-        new Position(Constants.ComplexPositionFEN),
-        new Position(Constants.KillerTestPositionFEN),
-    };
-
-    [Benchmark(Baseline = true)]
-    [ArgumentsSource(nameof(Data))]
-    public long Naive(Position position)
-    {
-        return CalculateCastleHash(position.Castle);
-    }
-
-    [Benchmark]
-    [ArgumentsSource(nameof(Data))]
-    public long Dictionary(Position position)
-    {
-        return ZobristTable.CastleHash(position.Castle);
-    }
-
-    [Benchmark]
-    [ArgumentsSource(nameof(Data))]
-    public long Switch(Position position)
-    {
-        return SwitchMethod(position);
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static long SwitchMethod(Position position)
-    {
-        return position.Castle switch
-        {
-            0 => 0,                                // -    | -
-
-            (byte)CastlingRights.WK => WK_Hash,    // K    | -
-            (byte)CastlingRights.WQ => WQ_Hash,    // Q    | -
-            (byte)CastlingRights.BK => BK_Hash,    // -    | k
-            (byte)CastlingRights.BQ => BQ_Hash,    // -    | q
-
-            (byte)CastlingRights.WK | (byte)CastlingRights.WQ => WK_Hash ^ WQ_Hash,    // KQ   | -
-            (byte)CastlingRights.WK | (byte)CastlingRights.BK => WK_Hash ^ BK_Hash,    // K    | k
-            (byte)CastlingRights.WK | (byte)CastlingRights.BQ => WK_Hash ^ BQ_Hash,    // K    | q
-            (byte)CastlingRights.WQ | (byte)CastlingRights.BK => WQ_Hash ^ BK_Hash,    // Q    | k
-            (byte)CastlingRights.WQ | (byte)CastlingRights.BQ => WQ_Hash ^ BQ_Hash,    // Q    | q
-            (byte)CastlingRights.BK | (byte)CastlingRights.BQ => BK_Hash ^ BQ_Hash,    // -    | kq
-
-            (byte)CastlingRights.WK | (byte)CastlingRights.WQ | (byte)CastlingRights.BK => WK_Hash ^ WQ_Hash ^ BK_Hash,    // KQ   | k
-            (byte)CastlingRights.WK | (byte)CastlingRights.WQ | (byte)CastlingRights.BQ => WK_Hash ^ WQ_Hash ^ BQ_Hash,    // KQ   | q
-            (byte)CastlingRights.WK | (byte)CastlingRights.BK | (byte)CastlingRights.BQ => WK_Hash ^ BK_Hash ^ BQ_Hash,    // K    | kq
-            (byte)CastlingRights.WQ | (byte)CastlingRights.BK | (byte)CastlingRights.BQ => WQ_Hash ^ BK_Hash ^ BQ_Hash,    // Q    | kq
-
-            (byte)CastlingRights.WK | (byte)CastlingRights.WQ | (byte)CastlingRights.BK | (byte)CastlingRights.BQ =>       // KQ   | kq
-                WK_Hash ^ WQ_Hash ^ BK_Hash ^ BQ_Hash,
-
-            _ => new()
-        };
-    }
-}
