+++ src/Lynx/ZobristTable.cs
@@ -11,6 +11,11 @@ public static class ZobristTable
 {
     private static readonly long[,] _table = Initialize();
 
+    private static readonly long WK_Hash = _table[(int)BoardSquare.a8, (int)Piece.p];
+    private static readonly long WQ_Hash = _table[(int)BoardSquare.b8, (int)Piece.p];
+    private static readonly long BK_Hash = _table[(int)BoardSquare.c8, (int)Piece.p];
+    private static readonly long BQ_Hash = _table[(int)BoardSquare.d8, (int)Piece.p];
+
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static long PieceHash(int boardSquare, int piece) => _table[boardSquare, piece];
 
@@ -49,35 +54,6 @@ public static class ZobristTable
         return _table[(int)BoardSquare.h8, (int)Piece.p];
     }
 
-    internal static readonly long WK_Hash = _table[(int)BoardSquare.a8, (int)Piece.p];
-    internal static readonly long WQ_Hash = _table[(int)BoardSquare.b8, (int)Piece.p];
-    internal static readonly long BK_Hash = _table[(int)BoardSquare.c8, (int)Piece.p];
-    internal static readonly long BQ_Hash = _table[(int)BoardSquare.d8, (int)Piece.p];
-
-    private static readonly Dictionary<byte, long> _castleHashDictionary = new()
-    {
-        [0] = 0,                                // -    | -
-        [(byte)CastlingRights.WK] = WK_Hash,    // K    | -
-        [(byte)CastlingRights.WQ] = WQ_Hash,    // Q    | -
-        [(byte)CastlingRights.BK] = BK_Hash,    // -    | k
-        [(byte)CastlingRights.BQ] = BQ_Hash,    // -    | q
-
-        [(byte)CastlingRights.WK | (byte)CastlingRights.WQ] = WK_Hash ^ WQ_Hash,    // KQ   | -
-        [(byte)CastlingRights.WK | (byte)CastlingRights.BK] = WK_Hash ^ BK_Hash,    // K    | k
-        [(byte)CastlingRights.WK | (byte)CastlingRights.BQ] = WK_Hash ^ BQ_Hash,    // K    | q
-        [(byte)CastlingRights.WQ | (byte)CastlingRights.BK] = WQ_Hash ^ BK_Hash,    // Q    | k
-        [(byte)CastlingRights.WQ | (byte)CastlingRights.BQ] = WQ_Hash ^ BQ_Hash,    // Q    | q
-        [(byte)CastlingRights.BK | (byte)CastlingRights.BQ] = BK_Hash ^ BQ_Hash,    // -    | kq
-
-        [(byte)CastlingRights.WK | (byte)CastlingRights.WQ | (byte)CastlingRights.BK] = WK_Hash ^ WQ_Hash ^ BK_Hash,    // KQ   | k
-        [(byte)CastlingRights.WK | (byte)CastlingRights.WQ | (byte)CastlingRights.BQ] = WK_Hash ^ WQ_Hash ^ BQ_Hash,    // KQ   | q
-        [(byte)CastlingRights.WK | (byte)CastlingRights.BK | (byte)CastlingRights.BQ] = WK_Hash ^ BK_Hash ^ BQ_Hash,    // K    | kq
-        [(byte)CastlingRights.WQ | (byte)CastlingRights.BK | (byte)CastlingRights.BQ] = WQ_Hash ^ BK_Hash ^ BQ_Hash,    // Q    | kq
-
-        [(byte)CastlingRights.WK | (byte)CastlingRights.WQ | (byte)CastlingRights.BK | (byte)CastlingRights.BQ] =       // KQ   | kq
-            WK_Hash ^ WQ_Hash ^ BK_Hash ^ BQ_Hash
-    };
-
     /// <summary>
     /// Uses <see cref="Piece.p"/> and
     /// <see cref="BoardSquare.a8"/> for <see cref="CastlingRights.WK"/>, <see cref="BoardSquare.b8"/> for <see cref="CastlingRights.WQ"/>
@@ -88,36 +64,32 @@ public static class ZobristTable
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static long CastleHash(byte castle)
     {
-        System.Diagnostics.Debug.Assert(_castleHashDictionary[castle] == CalculateCastleHash(castle));
-
-        return _castleHashDictionary[castle];
-    }
-
-    internal static long CalculateCastleHash(byte castle)
-    {
-        long combinedHash = 0;
-
-        if ((castle & (int)CastlingRights.WK) != default)
+        return castle switch
         {
-            combinedHash ^= _table[(int)BoardSquare.a8, (int)Piece.p];        // a8
-        }
+            0 => 0,                                // -    | -
 
-        if ((castle & (int)CastlingRights.WQ) != default)
-        {
-            combinedHash ^= _table[(int)BoardSquare.b8, (int)Piece.p];        // b8
-        }
+            (byte)CastlingRights.WK => WK_Hash,    // K    | -
+            (byte)CastlingRights.WQ => WQ_Hash,    // Q    | -
+            (byte)CastlingRights.BK => BK_Hash,    // -    | k
+            (byte)CastlingRights.BQ => BQ_Hash,    // -    | q
 
-        if ((castle & (int)CastlingRights.BK) != default)
-        {
-            combinedHash ^= _table[(int)BoardSquare.c8, (int)Piece.p];        // c8
-        }
+            (byte)CastlingRights.WK | (byte)CastlingRights.WQ => WK_Hash ^ WQ_Hash,    // KQ   | -
+            (byte)CastlingRights.WK | (byte)CastlingRights.BK => WK_Hash ^ BK_Hash,    // K    | k
+            (byte)CastlingRights.WK | (byte)CastlingRights.BQ => WK_Hash ^ BQ_Hash,    // K    | q
+            (byte)CastlingRights.WQ | (byte)CastlingRights.BK => WQ_Hash ^ BK_Hash,    // Q    | k
+            (byte)CastlingRights.WQ | (byte)CastlingRights.BQ => WQ_Hash ^ BQ_Hash,    // Q    | q
+            (byte)CastlingRights.BK | (byte)CastlingRights.BQ => BK_Hash ^ BQ_Hash,    // -    | kq
 
-        if ((castle & (int)CastlingRights.BQ) != default)
-        {
-            combinedHash ^= _table[(int)BoardSquare.d8, (int)Piece.p];        // d8
-        }
+            (byte)CastlingRights.WK | (byte)CastlingRights.WQ | (byte)CastlingRights.BK => WK_Hash ^ WQ_Hash ^ BK_Hash,    // KQ   | k
+            (byte)CastlingRights.WK | (byte)CastlingRights.WQ | (byte)CastlingRights.BQ => WK_Hash ^ WQ_Hash ^ BQ_Hash,    // KQ   | q
+            (byte)CastlingRights.WK | (byte)CastlingRights.BK | (byte)CastlingRights.BQ => WK_Hash ^ BK_Hash ^ BQ_Hash,    // K    | kq
+            (byte)CastlingRights.WQ | (byte)CastlingRights.BK | (byte)CastlingRights.BQ => WQ_Hash ^ BK_Hash ^ BQ_Hash,    // Q    | kq
 
-        return combinedHash;
+            (byte)CastlingRights.WK | (byte)CastlingRights.WQ | (byte)CastlingRights.BK | (byte)CastlingRights.BQ =>       // KQ   | kq
+                WK_Hash ^ WQ_Hash ^ BK_Hash ^ BQ_Hash,
+
+            _ => throw new($"Unexpected castle encoded number: {castle}")
+        };
     }
 
     /// <summary>
