+++ tests/Lynx.Test/ZobristTableTest.cs
@@ -96,6 +96,35 @@ public class ZobristTableTest
 
         var castleHash = ZobristTable.CastleHash(position.Castle);
 
+        Assert.AreEqual(CalculateCastleHash(position.Castle)castleHash);
+
         Assert.AreEqual(positionWithoutCastlingRightsHash, positionHash ^ castleHash);
+    }
+
+    private static long CalculateCastleHash(byte castle)
+    {
+        long combinedHash = 0;
+
+        if ((castle & (int)CastlingRights.WK) != default)
+        {
+            combinedHash ^= _zobristTable[(int)BoardSquare.a8, (int)Piece.p];        // a8
+        }
+
+        if ((castle & (int)CastlingRights.WQ) != default)
+        {
+            combinedHash ^= _zobristTable[(int)BoardSquare.b8, (int)Piece.p];        // b8
+        }
+
+        if ((castle & (int)CastlingRights.BK) != default)
+        {
+            combinedHash ^= _zobristTable[(int)BoardSquare.c8, (int)Piece.p];        // c8
+        }
+
+        if ((castle & (int)CastlingRights.BQ) != default)
+        {
+            combinedHash ^= _zobristTable[(int)BoardSquare.d8, (int)Piece.p];        // d8
+        }
+
+        return combinedHash;
     }
 }
